Add IAppVersionProvider; Extracted GitCommandExecutor out of AppVersionComposer
