+++ Assets/Scripts/Runtime/AppVersionHolder.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 namespace AppVersioning
 {
-    public partial class AppVersionHolder : ScriptableObject
+    public partial class AppVersionHolder : ScriptableObject, IAppVersionProvider
     {
         [SerializeField] private string _appVersion;
