+++ Assets/Scripts/Runtime/IAppVersionProvider.cs
@@ -1,0 +1,7 @@
+namespace AppVersioning
+{
+    public interface IAppVersionProvider
+    {
+        string AppVersion { get; }
+    }
+}
